Render show-more button
